GetCell関数はCellクラスに持たせる
